requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.2.10
astroid==2.3.3
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
code128==0.3
colorama==0.4.3
cryptography==3.0
cycler==0.10.0
DateTime==4.3
defusedxml==0.7.0rc1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.8
django-core==1.4.1
django-cors-headers==3.4.0
django-crispy-forms==1.8.1
django-filter==2.2.0
django-login-required-middleware==0.4
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
filelock==3.0.12
Flask==1.1.2
fpdf==1.7.2
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
img2pdf==0.2.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.4.1
mccabe==0.6.1
mysql-connector-python==8.0.23
mysqlclient==2.0.3
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.5.1.48
pdf2image==1.14.0
pdfkit==0.6.1
Pillow==7.2.0
protobuf==3.15.8
psycopg2==2.8.5
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
PyMuPDF==1.18.12
pyparsing==2.4.7
PyPDF2==1.26.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
python-barcode==0.13.1
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2020.1
qrcode==6.1
reportlab==3.5.66
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
Unipath==1.1
urllib3==1.25.10
virtualenv==20.4.3
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
wkhtmltopdf==0.2
wrapt==1.11.2
xhtml2pdf==0.2.5
zope.interface==5.1.0
